Segregate some search utils into their file
